discord cs2 bot/requirements.txt
discord.py  
python-dotenv